+++ history.md
@@ -1,6 +1,7 @@
 ## Edge
 
 * New: Add a throbber to indicate the project information is loading.
+* New: Display error message when pushing or pulling a language that doesn’t exist in a project.
 
 ## Version 2.0.3 / 2012-06-07